refactor(js-themes-toolkit): extract common code in upgrade process
